+++ FileService/src/FileService.Contracts/StartMultipartUploadRequest.cs
@@ -3,4 +3,4 @@
 public record StartMultipartUploadRequest(
     string FileName,
     string ContentType,
-    long Size);
+    int Size);
